Patch for code acquiring Let's Encrypt certificate.
We wanted let's encrypt certificate to be obtained automatically at launch.
Previously, config file for letsencrypt did not receive a value for DOMAIN.
This patch makes it so that the userdata, used to launch the nginx instance,
also declares a global variable DOMAIN once the instance is up and running.