Patch Go web server's user data.
Previously the biggest problem was how the paths would change, while in sudo and ubuntu.
Initially I tried to see if there's a way to reliably track the changes in relative path.
But it turned out the problem was that the directory created during userdata stage is run
as sudo so when user is logged in as ubuntu he doesn't have permissions to the work/ directory.
So all I had to do is change the ownership of the directory, and specify the $GOPATH.
One thing I also had to do though is export $GOPATH and $PATH in both the userdata script and also
cat it into profile.d/xxx.sh file so that when the process finishes, bashrc will set these env variables.

TO-DO:

Pull from bitbucket or something on launch.
Better way to sync s3 userdata files.
Configure webserver to interact with database.
Write database for basic user and order placements.
Hook it up with Stripe's payment API.
Clean up UI.